// source/frontend/src/models/api/generated/ApiGen_Mayan_DocumentMetadata.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Mayan_DocumentDetail } from './ApiGen_Mayan_DocumentDetail';
import { ApiGen_Mayan_MetadataType } from './ApiGen_Mayan_MetadataType';

// LINK: @backend/apimodels/Models/Mayan/Document/DocumentMetadataModel.cs
export interface ApiGen_Mayan_DocumentMetadata {
  document: ApiGen_Mayan_DocumentDetail | null;
  id: number;
  metadata_type: ApiGen_Mayan_MetadataType | null;
  url: string | null;
  value: string | null;
}